+++ internal/column/range_test.go
@@ -4,6 +4,7 @@ import (
 	"bytes"
 	"fmt"
 	"github.com/stretchr/testify/assert"
+	"github.com/xztaityozx/sel/internal/output"
 	"math/rand"
 	"strings"
 	"testing"
@@ -59,7 +60,7 @@ func TestRangeSelector_Select(t *testing.T) {
 		for _, v := range dataset {
 			rs := NewRangeSelector(v.start, v.step, v.stop, false)
 			expect := expectFactory(v.expects)
-			writer := NewWriter(" ", w)
+			writer := output.NewWriter(" ", w)
 			err := rs.Select(writer, &testEnumerable{a: cols})
 			assert.Nil(t, writer.Flush())
 			assert.Nil(t, err)
@@ -81,7 +82,7 @@ func TestRangeSelector_Select(t *testing.T) {
 
 		for _, v := range dataset {
 			rs := NewRangeSelector(v.start, v.step, v.stop, false)
-			writer := NewWriter(" ", w)
+			writer := output.NewWriter(" ", w)
 			err := rs.Select(writer, &testEnumerable{a: cols})
 			assert.Nil(t, writer.Flush())
 			assert.NotNil(t, err)
@@ -92,7 +93,7 @@ func TestRangeSelector_Select(t *testing.T) {
 
 	t.Run("Inf", func(t *testing.T) {
 		rs := NewRangeSelector(1, 1, 1, true)
-		writer := NewWriter(" ", w)
+		writer := output.NewWriter(" ", w)
 		err := rs.Select(writer, &testEnumerable{a: cols})
 		assert.Nil(t, writer.Flush())
 		assert.Nil(t, err)
