+++ internal/column/selector.go
@@ -2,8 +2,9 @@ package column
 
 import (
 	"github.com/xztaityozx/sel/internal/iterator"
+	"github.com/xztaityozx/sel/internal/output"
 )
 
 type Selector interface {
-	Select(w *Writer, iterator iterator.IEnumerable) error
+	Select(w *output.Writer, iterator iterator.IEnumerable) error
 }
