+++ internal/column/switch_test.go
@@ -4,6 +4,7 @@ import (
 	"bytes"
 	"github.com/stretchr/testify/assert"
 	"github.com/xztaityozx/sel/internal/iterator"
+	"github.com/xztaityozx/sel/internal/output"
 	"github.com/xztaityozx/sel/test"
 	"regexp"
 	"strings"
@@ -73,7 +74,7 @@ func TestSwitchSelector_Select(t *testing.T) {
 		end   endAddress
 	}
 	type args struct {
-		w    *Writer
+		w    *output.Writer
 		iter iterator.IEnumerable
 	}
 
@@ -98,7 +99,7 @@ func TestSwitchSelector_Select(t *testing.T) {
 				begin: address{num: 1},
 				end:   endAddress{address: address{num: 5}},
 			},
-			args: args{iter: &testEnumerable{a: cols}, w: NewWriter(" ", w)},
+			args: args{iter: &testEnumerable{a: cols}, w: output.NewWriter(" ", w)},
 			want: cols[0:5],
 		},
 		{
@@ -107,7 +108,7 @@ func TestSwitchSelector_Select(t *testing.T) {
 				begin: address{regexp: regexp.MustCompile(`a`)},
 				end:   endAddress{address{regexp: regexp.MustCompile(`e`)}, false},
 			},
-			args: args{iter: &testEnumerable{a: []string{"1", "2", "a", "b", "c", "d", "e", "3", "4"}}, w: NewWriter(" ", w)},
+			args: args{iter: &testEnumerable{a: []string{"1", "2", "a", "b", "c", "d", "e", "3", "4"}}, w: output.NewWriter(" ", w)},
 			want: []string{"a", "b", "c", "d", "e"},
 		},
 		{
@@ -116,7 +117,7 @@ func TestSwitchSelector_Select(t *testing.T) {
 				begin: address{regexp: regexp.MustCompile(`a`)},
 				end:   endAddress{address{num: 5}, true},
 			},
-			args: args{iter: &testEnumerable{a: []string{"1", "2", "a", "b", "c", "d", "e", "3", "4"}}, w: NewWriter(" ", w)},
+			args: args{iter: &testEnumerable{a: []string{"1", "2", "a", "b", "c", "d", "e", "3", "4"}}, w: output.NewWriter(" ", w)},
 			want: []string{"a", "b", "c", "d", "e", "3"},
 		},
 		{
@@ -125,7 +126,7 @@ func TestSwitchSelector_Select(t *testing.T) {
 				begin: address{regexp: regexp.MustCompile(`e`)},
 				end:   endAddress{address{num: -5}, true},
 			},
-			args: args{iter: &testEnumerable{a: []string{"1", "2", "a", "b", "c", "d", "e", "3", "4"}}, w: NewWriter(" ", w)},
+			args: args{iter: &testEnumerable{a: []string{"1", "2", "a", "b", "c", "d", "e", "3", "4"}}, w: output.NewWriter(" ", w)},
 			want: []string{"2", "a", "b", "c", "d", "e"},
 		},
 		{
@@ -134,7 +135,7 @@ func TestSwitchSelector_Select(t *testing.T) {
 				begin: address{regexp: regexp.MustCompile(`a`)},
 				end:   endAddress{address{num: -5}, true},
 			},
-			args: args{iter: &testEnumerable{a: []string{"1", "2", "a", "b", "c", "d", "e", "3", "4"}}, w: NewWriter(" ", w)},
+			args: args{iter: &testEnumerable{a: []string{"1", "2", "a", "b", "c", "d", "e", "3", "4"}}, w: output.NewWriter(" ", w)},
 			want: []string{"1", "2", "a"},
 		},
 		{
@@ -143,7 +144,7 @@ func TestSwitchSelector_Select(t *testing.T) {
 				begin: address{regexp: regexp.MustCompile(`e`)},
 				end:   endAddress{address{num: 5}, true},
 			},
-			args: args{iter: &testEnumerable{a: []string{"1", "2", "a", "b", "c", "d", "e", "3", "4"}}, w: NewWriter(" ", w)},
+			args: args{iter: &testEnumerable{a: []string{"1", "2", "a", "b", "c", "d", "e", "3", "4"}}, w: output.NewWriter(" ", w)},
 			want: []string{"e", "3", "4"},
 		},
 		{
@@ -152,7 +153,7 @@ func TestSwitchSelector_Select(t *testing.T) {
 				begin: address{regexp: regexp.MustCompile(`e`)},
 				end:   endAddress{address{num: 0}, true},
 			},
-			args: args{iter: &testEnumerable{a: []string{"1", "2", "a", "b", "c", "d", "e", "3", "4"}}, w: NewWriter(" ", w)},
+			args: args{iter: &testEnumerable{a: []string{"1", "2", "a", "b", "c", "d", "e", "3", "4"}}, w: output.NewWriter(" ", w)},
 			want: []string{"e"},
 		},
 		{
@@ -161,7 +162,7 @@ func TestSwitchSelector_Select(t *testing.T) {
 				begin: address{regexp: regexp.MustCompile(`e`)},
 				end:   endAddress{address{num: 0}, false},
 			},
-			args: args{iter: &testEnumerable{a: []string{"1", "2", "a", "b", "c", "d", "e", "3", "4"}}, w: NewWriter(" ", w)},
+			args: args{iter: &testEnumerable{a: []string{"1", "2", "a", "b", "c", "d", "e", "3", "4"}}, w: output.NewWriter(" ", w)},
 			want: []string{"e", "3", "4"},
 		},
 	}
