+++ internal/output/writer_test.go
@@ -1,4 +1,4 @@
-package column
+package output
 
 import (
 	"bufio"
